+++ cmake/BipedalLocomotionFrameworkFindDependencies.cmake
@@ -56,7 +56,7 @@ macro(checkandset_dependency package)
       find_package(${package} ${CSD_MINIMUM_VERSION} COMPONENTS ${CSD_COMPONENTS} REQUIRED)
     else ()
       find_package(${package} ${CSD_MINIMUM_VERSION} REQUIRED)
-      endif ()
+    endif ()
   endif()
 
   # FRAMEWORK_USE_SYSTEM_${package}
@@ -150,6 +150,18 @@ checkandset_dependency(cppad)
 find_package(manif QUIET)
 checkandset_dependency(manif)
 
+find_package(Python3 3.6 COMPONENTS Interpreter Development QUIET)
+checkandset_dependency(Python3 MINIMUM_VERSION 3.6 COMPONENTS Interpreter Development)
+
+find_package(pybind11 2.2 CONFIG QUIET)
+checkandset_dependency(pybind11)
+
+find_package(pytest QUIET)
+checkandset_dependency(pytest)
+
+find_package(matioCpp QUIET)
+checkandset_dependency(matioCpp)
+
 framework_dependent_option(FRAMEWORK_COMPILE_YarpUtilities
   "Compile YarpHelper library?" ON
   "FRAMEWORK_USE_YARP" OFF)
@@ -161,6 +173,10 @@ framework_dependent_option(FRAMEWORK_COMPILE_YarpImplementation
 framework_dependent_option(FRAMEWORK_COMPILE_Estimators
   "Compile Estimators library?" ON
   "" OFF)
+
+framework_dependent_option(FRAMEWORK_COMPILE_Contact
+  "Compile Contact libraries?" ON
+  "FRAMEWORK_USE_manif" OFF)
 
 framework_dependent_option(FRAMEWORK_COMPILE_Planners
   "Compile Planners libraries?" ON
@@ -186,3 +202,22 @@ framework_dependent_option(FRAMEWORK_COMPILE_FloatingBaseEstimators
   "Compile FloatingBaseEstimators libraries?" ON
   "FRAMEWORK_USE_manif" OFF)
 
+framework_dependent_option(FRAMEWORK_COMPILE_ManifConversions
+  "Compile manif Conversions libraries?" ON
+  "FRAMEWORK_USE_manif" OFF)
+
+framework_dependent_option(FRAMEWORK_COMPILE_matioCppConversions
+  "Compile matioCpp Conversions libraries?" ON
+  "FRAMEWORK_USE_matioCpp" OFF)
+
+framework_dependent_option(FRAMEWORK_COMPILE_JointPositionTrackingApplication
+  "Compile joint-position-tracking application?" ON
+  "FRAMEWORK_COMPILE_YarpImplementation;FRAMEWORK_COMPILE_Planners;FRAMEWORK_COMPILE_RobotInterface" OFF)
+
+framework_dependent_option(FRAMEWORK_COMPILE_PYTHON_BINDINGS
+  "Do you want to generate and compile the Python bindings?" ON
+  "FRAMEWORK_USE_Python3;FRAMEWORK_USE_pybind11" OFF)
+
+framework_dependent_option(FRAMEWORK_TEST_PYTHON_BINDINGS
+  "Do you want to test the Python bindings?" ON
+  "FRAMEWORK_COMPILE_tests;FRAMEWORK_COMPILE_PYTHON_BINDINGS;FRAMEWORK_USE_pytest" OFF)